+++ test/test_cppparser.cpp
@@ -57,6 +57,10 @@ TEST_CASE("c++ lexer" "[instancemgr]") {
 	CHECK(CPP_TEST_PASSED == test_lexer(
 		"struct SomeStruct {int a;}",
 		{CPP_KEYWORD, CPP_IDENTIFIER, CPP_OPERATOR, CPP_KEYWORD, CPP_IDENTIFIER, CPP_OPERATOR, CPP_OPERATOR}));
+	
+	CHECK(CPP_TEST_PASSED == test_lexer(
+		"enum SomeEnum {A=1}",
+		{CPP_KEYWORD, CPP_IDENTIFIER, CPP_OPERATOR, CPP_IDENTIFIER, CPP_OPERATOR, CPP_INTEGER_LITERAL, CPP_OPERATOR}));
 }
 
 static s32 test_lexer(const char* src, std::vector<CppTokenType>&& expected) {
@@ -152,6 +156,35 @@ TEST_CASE("c++ parser" "[instancemgr]") {
 		}()
 	));
 	CHECK(test_parser(
+		"struct /* comment */ SomeVars /* comment */ { s8 byte; s16 halfword; s32 word; s64 doubleword; s128 quadword; };",
+		[]() {
+			CppType type(CPP_STRUCT_OR_UNION);
+			
+			type.name = "SomeVars";
+			CppType& byte = type.struct_or_union.fields.emplace_back(CPP_BUILT_IN);
+			byte.name = "byte";
+			byte.built_in = CPP_S8;
+			
+			CppType& halfword = type.struct_or_union.fields.emplace_back(CPP_BUILT_IN);
+			halfword.name = "halfword";
+			halfword.built_in = CPP_S16;
+			
+			CppType& word = type.struct_or_union.fields.emplace_back(CPP_BUILT_IN);
+			word.name = "word";
+			word.built_in = CPP_S32;
+			
+			CppType& doubleword = type.struct_or_union.fields.emplace_back(CPP_BUILT_IN);
+			doubleword.name = "doubleword";
+			doubleword.built_in = CPP_S64;
+			
+			CppType& quadword = type.struct_or_union.fields.emplace_back(CPP_BUILT_IN);
+			quadword.name = "quadword";
+			quadword.built_in = CPP_S128;
+			
+			return type;
+		}()
+	));
+	CHECK(test_parser(
 		"union Union { float **double_pointer; };",
 		[]() {
 			CppType type(CPP_STRUCT_OR_UNION);
@@ -163,6 +196,17 @@ TEST_CASE("c++ parser" "[instancemgr]") {
 			CppType& inner = *field.pointer_or_reference.value_type.get();
 			inner.pointer_or_reference.value_type = std::make_unique<CppType>(CPP_BUILT_IN);
 			inner.pointer_or_reference.value_type->built_in = CPP_FLOAT;
+			return type;
+		}()
+	));
+	CHECK(test_parser(
+		"enum Enum { A = 1, B = 2, C = 3 };",
+		[]() {
+			CppType type(CPP_ENUM);
+			type.name = "Enum";
+			type.enumeration.constants.emplace_back(1, "A");
+			type.enumeration.constants.emplace_back(2, "B");
+			type.enumeration.constants.emplace_back(3, "C");
 			return type;
 		}()
 	));